+++ src/main/java/haage/hiddenarmour/client/HiddenArmourModMenu.java
@@ -3,7 +3,7 @@ package haage.hiddenarmour.client;
 import com.terraformersmc.modmenu.api.ConfigScreenFactory;
 import com.terraformersmc.modmenu.api.ModMenuApi;
 
-import net.minecraft.client.gui.screen.Screen;
+import net.minecraft.client.gui.screens.Screen;
 
 public class HiddenArmourModMenu implements ModMenuApi {
     @Override