+++ src/main/java/haage/hiddenarmour/mixin/EquipmentModelMixin.java
@@ -6,11 +6,11 @@ import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 import haage.hiddenarmour.config.HiddenArmourConfig;
-import net.minecraft.item.ItemStack;
+import net.minecraft.world.item.ItemStack;
 
 @Mixin(ItemStack.class)
 public class EquipmentModelMixin {
-    @Inject(method = "hasGlint", at = @At("HEAD"), cancellable = true)
+    @Inject(method = "hasFoil", at = @At("HEAD"), cancellable = true)
     private void disableGlint(CallbackInfoReturnable<Boolean> cir) {
         if (HiddenArmourConfig.get().hideEnchantmentGlint) {
             cir.setReturnValue(false);